+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # Unreleased
 
+- Update dependencies, mainly mocha. The min nodejs version is now 20.19.0.
+  In mocha tests, use of `only()` is now blocked by default: remove them or use --no-forbid-only`.
+
 # 0.37.0 2026-03-24
 
 - findDoubleProperty and findIntProperty added to PropertyWrapper